Added BindToRenderStep function
Added a function that allows for a render step function to be easily added to a janitor.

Currently, if you wish to create add a render step bind to the janitor, you must just add the Unbind function within a custom function. This simplifies that process and does it all in one go.